✏️ Refactor Thunks #103 - Refactor Part 3
